+++ .idea/encodings.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="Encoding">
+    <file url="file://$PROJECT_DIR$/FactionsModule/src/main/java" charset="UTF-8" />
+    <file url="file://$PROJECT_DIR$/FactionsModule/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/NexusCore/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/NexusCore/src/main/resources" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/NexusFactionsAddon/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/NexusFactionsAddon/src/main/resources" charset="UTF-8" />
   </component>
 </project>
