+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/FactionsModule/FactionsModule.iml" filepath="$PROJECT_DIR$/FactionsModule/FactionsModule.iml" />
       <module fileurl="file://$PROJECT_DIR$/NexusCore/NexusCore.iml" filepath="$PROJECT_DIR$/NexusCore/NexusCore.iml" />
-      <module fileurl="file://$PROJECT_DIR$/NexusFactionsAddon/NexusFactionsAddon.iml" filepath="$PROJECT_DIR$/NexusFactionsAddon/NexusFactionsAddon.iml" />
     </modules>
   </component>
 </project>
