+++ FactionsModule/src/main/java/brawl/factionsmodule/util/FactionOperations.java
@@ -1,6 +1,6 @@
-package util;
+package brawl.factionsmodule.util;
 
-import brawl.factionsnexus.FactionsNexusController;
+import brawl.factionsmodule.FactionsNexusController;
 import brawl.nexuscore.NexusController;
 import com.massivecraft.factions.Faction;
 import main.java.tasks.FactionRefreshBarrierTask;
