+++ FactionsModule/src/main/resources/config.yml
