+++ FactionsModule/src/main/resources/plugin.yml
