--- NexusFactionsAddon/src/main/java/brawl/factionsmodule/FactionsModule.java
@@ -1,17 +1,0 @@
-package brawl.factionsmodule;
-
-import org.bukkit.plugin.java.JavaPlugin;
-
-public final class FactionsModule extends JavaPlugin {
-
-    @Override
-    public void onEnable() {
-        // Plugin startup logic
-
-    }
-
-    @Override
-    public void onDisable() {
-        // Plugin shutdown logic
-    }
-}
